Refactor. feature toggle
