+++ speech-asistant-proxy/default.conf
@@ -13,10 +13,14 @@ server {
         try_files $uri $uri/ =404;
     }
 
-    location /speech-assistant/feature {
-       add_header content-type "text/plain"; 
+    location /speech-assistant/update/feature {
+        default_type text/plain;
         js_content featureToggle.update;
     }
 
+    location /speech-assistant/get/feature {
+        default_type text/plain;
+        js_content featureToggle.get;
+    }
     
 }
