+++ speech-asistant-proxy/nginx.conf
