+++ speech-asistant-proxy/speech-assistant/feature-enable.txt
